refactor(cli/migrate): rename migration 'update-remix-imports' to 'replace-remix-imports'
