+++ packages/remix-dev/cli/migrate/migration-options.ts
@@ -1,6 +1,6 @@
 export const migrationOptions = [
   {
-    name: "update-remix-imports: Replaces `remix` package import statements with specific `@remix-run/*` package import statements.",
-    value: "update-remix-imports",
+    name: "replace-remix-imports: Replaces `remix` package import statements with specific `@remix-run/*` package import statements.",
+    value: "replace-remix-imports",
   },
 ] as const;
