+++ packages/remix-dev/cli/migrate/run/transforms/index.ts
@@ -1,2 +1,2 @@
 export * from "./types";
-export * from "./update-remix-imports";
+export * from "./replace-remix-imports";
