+++ packages/remix-dev/cli/migrate/run/transforms/replace-remix-imports/cleanup-package-json.ts
